add graphiql playground into dashboard
