+++ frontend/package.json
@@ -76,8 +76,8 @@
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "file-loader": "^6.2.0",
+    "graphiql": "^1.2.0",
     "graphql": "^15.4.0",
-    "graphql-playground-react": "^1.7.27",
     "isomorphic-fetch": "^3.0.0",
     "jsonwebtoken": "^8.5.1",
     "moment": "^2.29.1",
