--- frontend/src/app/dashboard/components/Playground/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Playground'
